src/requirements.txt
Flask>=1.1.0
clamd
azure-storage-blob